process asyncronously
