+++ custom/icds/location_reassignment/utils.py
@@ -6,12 +6,6 @@ from casexml.apps.case.mock import CaseBlock
 from corehq.apps.hqcase.utils import submit_case_blocks
 from corehq.apps.locations.models import SQLLocation
 from corehq.apps.locations.tasks import deactivate_users_at_location
-from corehq.apps.userreports.data_source_providers import (
-    DynamicDataSourceProvider,
-    StaticDataSourceProvider,
-)
-from corehq.apps.userreports.specs import EvaluationContext
-from corehq.apps.userreports.util import get_indicator_adapter
 from corehq.apps.users.util import SYSTEM_USER_ID
 from corehq.form_processor.backends.sql.dbaccessors import CaseAccessorSQL
 from custom.icds.location_reassignment.const import (
@@ -19,6 +13,7 @@ from custom.icds.location_reassignment.const import (
     HOUSEHOLD_CASE_TYPE,
 )
 from custom.icds.location_reassignment.models import Transition
+from custom.icds.location_reassignment.tasks import process_ucr_changes
 
 
 def deprecate_locations(domain, old_locations, new_locations, operation):
@@ -107,33 +102,4 @@ def reassign_household_case(domain, household_case_id, old_owner_id, new_owner_i
         case_blocks.append(case_block)
     if case_blocks:
         submit_case_blocks(case_blocks, domain, user_id=SYSTEM_USER_ID)
-        process_ucr_changes(domain, case_ids)
-
-
-def process_ucr_changes(domain, case_ids):
-    cases = CaseAccessorSQL.get_cases(case_ids)
-    docs = [case.to_json() for case in cases]
-    data_source_providers = [DynamicDataSourceProvider(), StaticDataSourceProvider()]
-
-    all_configs = [
-        source
-        for provider in data_source_providers
-        for source in provider.by_domain(domain)
-    ]
-
-    adapters = [
-        get_indicator_adapter(config, raise_errors=True, load_source='location_reassignment')
-        for config in all_configs
-    ]
-
-    async_configs_by_doc_id = {}
-    for doc in docs:
-        eval_context = EvaluationContext(doc)
-        for adapter in adapters:
-            if adapter.config.filter(doc, eval_context):
-                async_configs_by_doc_id[doc['_id']].append(adapter.config._id)
-                rows_to_save = adapter.get_all_values(doc, eval_context)
-                if rows_to_save:
-                    adapter.save_rows(rows_to_save, reassigning_cases=True)
-                else:
-                    adapter.delete(doc, reassigning_cases=True)
+        process_ucr_changes.delay(domain, case_ids)
